Just fixed some errors and unclear phrasing. (#201)
I'm still a little unsure of what  Line 48: `If using a KMS it
will make it easy to globally change across numerous Lambda's the location of
a daemon, the log level or the license key if you need to change where the data
reports.` means. It seems like it's saying "If you're using a KMS and you set up your configuration to use ENVIRONMENT VARIABLES, it will be easier to globally change the location of the daemon, the log level, or the license key (if you need to change where the data reports) across multiple Lambdas.  

I'm not a SME so I'm not 100% sure what it is supposed to say, but it should maybe be clarified.

I work at New Relic - DJ Zevenbergen